extract_data/lyrics/obtener_canciones/requirements.txt
async-timeout==5.0.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
zstandard==0.23.0
